refactor: move MQSC command methods to commands module
